Fix button press leads to "undefined" value
There was this bug that lead to printing "undefined" to the display. It
was caused due to the fact that clicks were registered from the divs
which were enclosing the buttons i.e. ".row-1", ".row-2", etc, because
the querySelector was applied on ".buttons" div so that it would capture
events upon bubbling two levels up. Now, the event listeners are applied
on the buttons themselves so this minor possibility of their parent divs
getting clicked and registered is diminished.