Allow limited number of backspaces
Clear button will not function once the display reaches an operator. This
is due to the fact that we will need to empty the numbersArray and
operators array and repopulate it which is a hectic task. Also, a lot of
other problems also do arise, so it's better not to include that feature
for the time being.